feat: extract subject detail composable
